+++ .pnp.cjs
@@ -19,8 +19,8 @@ const RAW_RUNTIME_STATE =
       "reference": "workspace:packages/idsdb"\
     },\
     {\
-      "name": "@mandel59/idsdb-d1",\
-      "reference": "workspace:packages/idsdb-d1"\
+      "name": "@mandel59/idsdb-fts5",\
+      "reference": "workspace:packages/idsdb-fts5"\
     },\
     {\
       "name": "@mandel59/idsdb-utils",\
@@ -84,7 +84,7 @@ const RAW_RUNTIME_STATE =
   "pnpZipBackend": "libzip",\
   "fallbackExclusionList": [\
     ["@mandel59/idsdb", ["workspace:packages/idsdb"]],\
-    ["@mandel59/idsdb-d1", ["workspace:packages/idsdb-d1"]],\
+    ["@mandel59/idsdb-fts5", ["workspace:packages/idsdb-fts5"]],\
     ["@mandel59/idsdb-utils", ["workspace:packages/idsdb-utils"]],\
     ["@mandel59/idstool", ["workspace:packages/idstool"]],\
     ["@mandel59/mojidata", ["workspace:packages/mojidata"]],\
@@ -1072,16 +1072,17 @@ const RAW_RUNTIME_STATE =
           ["@types/node", "npm:24.10.4"],\
           ["better-sqlite3", "npm:12.5.0"],\
           ["ts-node", "virtual:df7814ce6017719e7a429bd7f5b57f20aecfa75188ccb4258d5d658150873d1a0016cce412ce65b7e3b1f96a5de22d0e7a145c0ad80f2645371336c9035c251a#npm:10.9.2"],\
+          ["tsx", "npm:4.21.0"],\
           ["typescript", "patch:typescript@npm%3A5.9.3#optional!builtin<compat/typescript>::version=5.9.3&hash=5786d5"]\
         ],\
         "linkType": "SOFT"\
       }]\
     ]],\
-    ["@mandel59/idsdb-d1", [\
-      ["workspace:packages/idsdb-d1", {\
-        "packageLocation": "./packages/idsdb-d1/",\
+    ["@mandel59/idsdb-fts5", [\
+      ["workspace:packages/idsdb-fts5", {\
+        "packageLocation": "./packages/idsdb-fts5/",\
         "packageDependencies": [\
-          ["@mandel59/idsdb-d1", "workspace:packages/idsdb-d1"]\
+          ["@mandel59/idsdb-fts5", "workspace:packages/idsdb-fts5"]\
         ],\
         "linkType": "SOFT"\
       }]\
@@ -1189,7 +1190,7 @@ const RAW_RUNTIME_STATE =
       ["virtual:786ff3bbd6eef7688a833439a338e46d6a1c41d5fd70226af8748ed2e6ea5e91db2c741160011bc5d1947617ca1a1682e3ff4cb4aa28f74212bb5b59b725f144#workspace:packages/mojidata-api-better-sqlite3", {\
         "packageLocation": "./.yarn/__virtual__/@mandel59-mojidata-api-better-sqlite3-virtual-9374ea863f/1/packages/mojidata-api-better-sqlite3/",\
         "packageDependencies": [\
-          ["@mandel59/idsdb", "workspace:packages/idsdb"],\
+          ["@mandel59/idsdb-fts5", "workspace:packages/idsdb-fts5"],\
           ["@mandel59/mojidata", "workspace:packages/mojidata"],\
           ["@mandel59/mojidata-api-better-sqlite3", "virtual:786ff3bbd6eef7688a833439a338e46d6a1c41d5fd70226af8748ed2e6ea5e91db2c741160011bc5d1947617ca1a1682e3ff4cb4aa28f74212bb5b59b725f144#workspace:packages/mojidata-api-better-sqlite3"],\
           ["@mandel59/mojidata-api-core", "workspace:packages/mojidata-api-core"],\
@@ -1208,7 +1209,7 @@ const RAW_RUNTIME_STATE =
       ["workspace:packages/mojidata-api-better-sqlite3", {\
         "packageLocation": "./packages/mojidata-api-better-sqlite3/",\
         "packageDependencies": [\
-          ["@mandel59/idsdb", "workspace:packages/idsdb"],\
+          ["@mandel59/idsdb-fts5", "workspace:packages/idsdb-fts5"],\
           ["@mandel59/mojidata", "workspace:packages/mojidata"],\
           ["@mandel59/mojidata-api-better-sqlite3", "workspace:packages/mojidata-api-better-sqlite3"],\
           ["@mandel59/mojidata-api-core", "workspace:packages/mojidata-api-core"],\
@@ -1265,7 +1266,7 @@ const RAW_RUNTIME_STATE =
       ["workspace:packages/mojidata-api-node-sqlite", {\
         "packageLocation": "./packages/mojidata-api-node-sqlite/",\
         "packageDependencies": [\
-          ["@mandel59/idsdb", "workspace:packages/idsdb"],\
+          ["@mandel59/idsdb-fts5", "workspace:packages/idsdb-fts5"],\
           ["@mandel59/mojidata", "workspace:packages/mojidata"],\
           ["@mandel59/mojidata-api-core", "workspace:packages/mojidata-api-core"],\
           ["@mandel59/mojidata-api-hono", "workspace:packages/mojidata-api-hono"],\