+++ packages/idsdb-fts5/CHANGELOG.md
@@ -1,4 +1,4 @@
-# Change Log
+# @mandel59/idsdb-fts5
 
 All notable changes to this project will be documented in this file.
 
@@ -9,6 +9,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 
-- Publish an FTS5-oriented `idsfind.db` package for Cloudflare D1 import flows.
+- Publish an FTS5-oriented `idsfind.db` package for mojidata backends that support SQLite FTS5.
 
 [1.8.0]: https://github.com/mandel59/mojidata/releases/tag/v1.8.0